+++ packages/robo/src/core/constants.ts
@@ -1,5 +1,8 @@
 import type { Config } from '../types/index.js'
 
+// TODO: Restore support for ['.js', '.jsx', '.ts', '.tsx']
+export const ALLOWED_EXTENSIONS = ['.js']
+
 export const DEFAULT_CONFIG: Config = {
 	clientOptions: null,
 	heartbeat: {